Add summarization microservices: implement content fetching, keyword extraction, and article processing modules
