+++ backend/microservices/summarization/__init__.py
@@ -1,0 +1,1 @@
+# This file marks the directory as a Python package
