+++ backend/microservices/summarization_service.py
@@ -14,12 +14,15 @@ Key Features:
 
 import json
 import requests
-from bs4 import BeautifulSoup
 import openai
 from backend.core.config import Config
 from backend.core.utils import setup_logger, log_exception
-import yake
 import os
+
+# Import the refactored modules
+from backend.microservices.summarization.content_fetcher import fetch_article_content
+from backend.microservices.summarization.keyword_extractor import get_keywords
+from backend.microservices.summarization.article_processor import process_articles
 
 # Initialize logger
 logger = setup_logger(__name__)
@@ -38,56 +41,6 @@ load_dotenv('../../.env')  # Optional: Only use this for local development.
 SUPABASE_URL = os.getenv("VITE_SUPABASE_URL")
 SUPABASE_SERVICE_KEY = os.getenv("VITE_SUPABASE_ANON_KEY")
 supabase: Client = create_client(SUPABASE_URL, SUPABASE_SERVICE_KEY)
-
-
-@log_exception(logger)
-def fetch_article_content(url):
-    """
-    Fetches and extracts the main content from a given URL.
-
-    Args:
-        url (str): The URL of the article to fetch content from.
-
-    Returns:
-        str or None: The extracted article content as plain text.
-                    Returns None if the fetch fails or content is invalid.
-    """
-    try:
-        if not url or not url.startswith('http'):
-            logger.error(f"Invalid URL format: {url}")
-            return None
-
-        headers = {
-            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
-        }
-        response = requests.get(url, headers=headers, timeout=10)
-        response.raise_for_status()
-        
-        soup = BeautifulSoup(response.text, 'html.parser')
-        paragraphs = soup.find_all('p')
-        
-        if not paragraphs:
-            logger.warning(f"No content found at URL: {url}")
-            return None
-            
-        content = ' '.join([p.get_text() for p in paragraphs])
-        return content
-        
-    except requests.exceptions.Timeout:
-        logger.error(f"Request timed out for URL: {url}")
-        return None
-    except requests.exceptions.SSLError:
-        logger.error(f"SSL verification failed for URL: {url}")
-        return None
-    except requests.exceptions.ConnectionError:
-        logger.error(f"Failed to connect to URL: {url}")
-        return None
-    except requests.exceptions.RequestException as e:
-        logger.error(f"Error fetching article content from {url}: {str(e)}")
-        return None
-    except Exception as e:
-        logger.error(f"Unexpected error processing {url}: {str(e)}")
-        return None
 
 
 @log_exception(logger)
@@ -121,124 +74,6 @@ def run_summarization(text):
     except Exception as e:
         logger.error(f"Error in summarization: {str(e)}")
         return "Error generating summary"
-
-
-@log_exception(logger)
-def get_keywords(text, num_keywords=1):
-    """
-    Extracts key phrases from the input text using YAKE keyword extraction.
-
-    Args:
-        text (str): The input text to extract keywords from.
-        num_keywords (int, optional): Number of keywords to extract. Defaults to 1.
-
-    Returns:
-        list: A list of extracted keywords/key phrases.
-    """
-    kw_extractor = yake.KeywordExtractor(top=num_keywords, lan='en')
-    keywords = kw_extractor.extract_keywords(text)
-    return [kw[0] for kw in keywords]
-
-
-@log_exception(logger)
-def process_articles(article_ids,user_id):
-    """
-    Processes a batch of articles associated with a specific session ID.
-    
-    This function performs the following operations:
-    1. Retrieves articles from Supabase based on the session ID.
-    2. Fetches missing content for articles if needed.
-    3. Generates summaries for each article.
-    4. Extracts keywords for filtering.
-    
-    Args:
-        article_ids (list): A list of article IDs to process.
-
-    Returns:
-        list: A list of dictionaries containing processed article data.
-    """
-    try:
-        # history_result = supabase.table("user_search_history").select("news_id").eq("session_id", session_id).execute()
-        # article_ids = [record["news_id"] for record in history_result.data]
-
-        # articles = []
-        # if article_ids:
-        #     result = supabase.table("news_articles").select("*").in_("id", article_ids).execute()
-        #     articles = result.data
-        
-
-        articles = []
-
-        # Step 1: Fetch the news_ids from user_bookmarks for the given user_id
-        bookmark_result = supabase.table("user_bookmarks").select("news_id").eq("user_id", user_id).execute()
-        bookmark_result = supabase.table("user_bookmarks").select("id, news_id").eq("user_id", user_id).execute()
-
-        bookmark_records = {}
-        if bookmark_result.data:
-            bookmark_records = {item["news_id"]: item["id"] for item in bookmark_result.data}
-
-
-        bookmarked_news_ids = set(item["news_id"] for item in bookmark_result.data) if bookmark_result.data else set()
-
-        print(f"Bookmarked news IDs: {bookmarked_news_ids}")
-        print(f"Article IDs: {article_ids}")
-
-        # Step 2: Fetch all articles from news_articles using the article_ids
-        if article_ids:  # Assuming article_ids is defined or fetched earlier
-            result = supabase.table("news_articles").select("*").in_("id", article_ids).execute()
-            articles = result.data
-
-        # Step 3: Add the 'bookmarked' key to each article
-        for article in articles:
-            # article["bookmarked_id"] = article["id"] if article["id"] in bookmarked_news_ids else None
-            article["bookmarked_id"] = bookmark_records.get(article["id"], None)
-
-        # # If article_ids isn't defined earlier, you can fetch it here as well
-        # if not article_ids and bookmark_result.data:
-        #     article_ids = [item["news_id"] for item in bookmark_result.data]
-        #     result = supabase.table("news_articles").select("*").in_("id", article_ids).execute()
-        #     articles = result.data
-        #     for article in articles:
-        #         article["bookmarked"] = "yes"  # All articles here are bookmarked
-        
-
-
-
-      
-        print(articles)
-
-        summarized_articles = []
-        for article in articles:
-            logger.info(f"Processing article: {article['title']}")
-            
-            content = article.get('content')
-            if not content:
-                content = fetch_article_content(article['url'])
-            
-            if content:
-                summary = run_summarization(content)
-            else:
-                summary = run_summarization(article.get('content', ''))
-            
-            summarized_articles.append({
-                'id': article['id'],
-                'title': article['title'],
-                'author': article.get('author', 'Unknown Author'),
-                'source': article.get('source'),
-                'publishedAt': article.get('published_at'),
-                'url': article['url'],
-                'urlToImage': article.get('image'),
-                'content': article.get('content', ''),
-                'summary': summary,
-                'filter_keywords': get_keywords(article.get('content', '')),
-                'bookmarked_id': article.get('bookmarked_id', None)
-            })
-
-        return summarized_articles
-
-    except Exception as e:
-        logger.error(f"Error processing articles: {str(e)}")
-        raise e
 
 
 if __name__ == '__main__':
